Import of JCROMIE/Log-Log4perl-AutoCategorize-0.03_51 from CPAN.
gitpan-cpan-distribution: Log-Log4perl-AutoCategorize
gitpan-cpan-version:      0.03_51
gitpan-cpan-path:         JCROMIE/Log-Log4perl-AutoCategorize-0.03_51.tar.gz
gitpan-cpan-author:       JCROMIE
gitpan-cpan-maturity:     developer

+++ t/99_cleanup.t
@@ -7,7 +7,8 @@ BEGIN {
 }
 use Test::More tests => 1;
 
-diag "deleting test output files: ". `ls out.*`;
-system 'rm out.*';
+diag "deleting test output files: ". `ls out.*`	if $ENV{TEST_VERBOSE};
+
+system 'rm out.*' unless $ENV{TEST_POSTMORTEM};
 
 ok(1);
